Paramaterize number of arguments disallowed.
