+++ lib/config/configuration.js
@@ -1122,7 +1122,7 @@ Configuration.prototype.registerDefaultRules = function() {
 
     this.registerRule(require('../rules/require-aligned-multiline-params'));
 
-    this.registerRule(require('../rules/disallow-spaces-inside-function-calls-with-single-argument'));
+    this.registerRule(require('../rules/disallow-spaces-inside-function-calls'));
 };
 
 /**
